refactored attendee section
