+++ src/components/booking/attendeeBookingList.js
@@ -1,0 +1,65 @@
+import React from 'react'
+import { AttendeeCheckboxItem } from './attendeeCheckboxItem'
+import { useStoreItems } from 'SVHooks/store/useStoreItems'
+import { useCurrentSession } from 'SVHooks/booking/useCurrentSession'
+import { useBookingSet } from 'SVHooks/booking/useBookingSet'
+import { useWaitingSet } from 'SVHooks/booking/useWaitingSet'
+import { useRestrictedAttendeeIds } from 'SVHooks/booking/useRestrictedAttendeeIds'
+
+/**
+ * Gets computed values about the state of all checkboxees in the attendee list
+ * @param {Object} session
+ * @returns {Object} { enableCheck }
+ */
+const useCheckboxState = session => {
+  const groupBookingCapacity = useStoreItems('groupBooking.capacity')
+  return {
+    enableCheck:
+      session?.capacity?.isUnlimited ||
+      session?.capacity?.isWaitingListAvailable ||
+      groupBookingCapacity > 0,
+  }
+}
+
+/**
+ * List of attendees for a group booking section
+ * @param {Object} props
+ * @param {Array<Attendee>} props.attendees - attendee list for current section
+ * @param {Object} props.itemStyles
+ * @param {Object} props.sectionStyles
+ * @param {Function} props.onAttendeeSelected
+ */
+export const AttendeeBookingList = ({
+  attendees,
+  itemStyles,
+  sectionStyles,
+  onAttendeeSelected,
+}) => {
+  const bookingList = useBookingSet()
+  const waitingList = useWaitingSet()
+  const session = useCurrentSession()
+
+  // get the isBookable callback to check if an attendee is eligible to book the session
+  const { isBookable } = useRestrictedAttendeeIds(session?.identifier)
+  const { enableCheck } = useCheckboxState(session)
+
+  return attendees?.map(({ bookedTicketIdentifier: attendeeId, name }) => {
+    const isBooking = bookingList.has(attendeeId)
+    const isWaiting = waitingList.has(attendeeId)
+
+    return (
+      <AttendeeCheckboxItem
+        key={attendeeId}
+        id={attendeeId}
+        name={name}
+        onAttendeeSelected={onAttendeeSelected}
+        isWaiting={isWaiting}
+        sectionStyles={sectionStyles}
+        itemStyles={itemStyles}
+        disabled={!isBookable?.(attendeeId)}
+        enableCheck={enableCheck}
+        checked={isBooking || isWaiting}
+      />
+    )
+  })
+}
