+++ src/components/booking/groupBookingSection.js
@@ -1,13 +1,9 @@
 import React, { useMemo } from 'react'
 import { Text } from '@keg-hub/keg-components'
+import { AttendeeBookingList } from './attendeeBookingList'
 import { CheckboxGroup } from 'SVComponents/group/checkboxGroup'
-import { useStylesMemo } from 'SVHooks/useStylesMemo'
-import { AttendeeCheckboxItem } from './attendeeCheckboxItem'
-import { useCurrentSession } from 'SVHooks/booking/useCurrentSession'
-import { useBookingSet } from 'SVHooks/booking/useBookingSet'
-import { useWaitingSet } from 'SVHooks/booking/useWaitingSet'
+import { useStylesMemo } from '@keg-hub/re-theme'
 import { useStoreItems } from 'SVHooks/store/useStoreItems'
-import { useRestrictedAttendeeIds } from 'SVHooks/booking/useRestrictedAttendeeIds'
 
 /**
  * Gets the list of Attendee objects for the current section,
@@ -39,19 +35,6 @@ export const GroupBookingSection = ({
     sectionStyles?.content?.item
   )
 
-  const bookingList = useBookingSet()
-  const waitingList = useWaitingSet()
-  const session = useCurrentSession()
-  const groupBookingCapacity = useStoreItems('groupBooking.capacity')
-
-  // get the isBookable callback to check if an attendee is eligible to book the session
-  const { isBookable } = useRestrictedAttendeeIds(session?.identifier)
-
-  const enableCheck =
-    session?.capacity?.isUnlimited ||
-    session?.capacity?.isWaitingListAvailable ||
-    groupBookingCapacity > 0
-
   const attendeesForSection = useSectionAttendees(attendeeIds)
 
   return (
@@ -59,27 +42,12 @@ export const GroupBookingSection = ({
       styles={sectionStyles}
       title={name}
     >
-      { attendeesForSection?.map(
-        ({ bookedTicketIdentifier: attendeeId, name }) => {
-          const isBooking = bookingList.has(attendeeId)
-          const isWaiting = waitingList.has(attendeeId)
-
-          return (
-            <AttendeeCheckboxItem
-              key={attendeeId}
-              id={attendeeId}
-              name={name}
-              onAttendeeSelected={onAttendeeSelected}
-              isWaiting={isWaiting}
-              sectionStyles={sectionStyles}
-              itemStyles={itemStyles}
-              disabled={!isBookable?.(attendeeId)}
-              enableCheck={enableCheck}
-              checked={isBooking || isWaiting}
-            />
-          )
-        }
-      ) }
+      <AttendeeBookingList
+        attendees={attendeesForSection}
+        onAttendeeSelected={onAttendeeSelected}
+        itemStyles={itemStyles}
+        sectionStyles={sectionStyles}
+      />
       { !attendeesForSection?.length && (
         <Text>No attendees for this category</Text>
       ) }
