--- src/hooks/useStylesMemo.js
@@ -1,14 +1,0 @@
-import { useMemo } from 'react'
-import { useTheme } from '@keg-hub/re-theme'
-
-const noOpObj = {
-  content: {},
-}
-
-export const useStylesMemo = (...styleArgs) => {
-  const theme = useTheme()
-  return useMemo(() => theme.get(...styleArgs) || noOpObj, [
-    theme,
-    ...styleArgs,
-  ])
-}
